Fix warnings with g++-4.1 in Scribo.
	* scribo/binarization/internal/first_pass_functor.hh,
	* scribo/binarization/sauvola_ms.hh,
	* scribo/binarization/sauvola_threshold_image.hh,
	* scribo/binarization/sauvola_threshold_image_debug.hh,
	* scribo/postprocessing/fill_object_holes.hh,
	* scribo/preprocessing/deskew.hh,
	* scribo/primitive/extract/lines_pattern.hh,
	* scribo/subsampling/bilinear.hh,
	* scribo/subsampling/integral_single_image.hh,
	* scribo/text/clean_inplace.hh,
	* src/debug/show_links_several_right_overlap.cc,
	* src/debug/show_links_single_down.cc,
	* src/debug/show_links_single_down_left_aligned.cc,
	* src/debug/show_links_single_down_right_aligned.cc,
	* src/debug/show_links_single_left.cc,
	* src/debug/show_links_single_right.cc,
	* src/debug/show_links_single_up.cc,
	* src/debug/show_links_single_up_left_aligned.cc,
	* src/debug/show_links_single_up_right_aligned.cc: explicitly
	convert to a specific type when necessary.

+++ scribo/scribo/binarization/sauvola_ms.hh
@@ -150,7 +150,7 @@ namespace scribo
 	typedef point2d P;
 
 	// Cast to float is needed on MacOS X.
-	unsigned ratio = std::pow(float(q), float(i - 2u));  // Ratio in comparison to e_2
+	unsigned ratio = unsigned(std::pow(float(q), float(i - 2u)));  // Ratio in comparison to e_2
 
 	unsigned
 	  w_local = w * ratio,
@@ -875,7 +875,7 @@ namespace scribo
 	  {
 	    int i = sub_ima.size() - 1;
 	    // Cast to float is needed on MacOS X.
-	    unsigned ratio = std::pow(float(q), float(i - 2)); // Ratio compared to e_2
+	    unsigned ratio = unsigned(std::pow(float(q), float(i - 2))); // Ratio compared to e_2
 	    t_ima[i] = internal::compute_t_n_and_e_2(sub_ima[i], e_2,
 						     lambda_min_2 / ratio,
 						     mln_max(unsigned),
@@ -890,7 +890,7 @@ namespace scribo
 	    for (int i = sub_ima.size() - 2; i > 2; --i)
 	    {
 	      // Cast to float is needed on MacOS X.
-	      unsigned ratio = std::pow(float(q), float(i - 2)); // Ratio compared to e_2
+	      unsigned ratio = unsigned(std::pow(float(q), float(i - 2))); // Ratio compared to e_2
 	      t_ima[i] = internal::compute_t_n_and_e_2(sub_ima[i], e_2,
 						       lambda_min_2 / ratio,
 						       lambda_max_2 / ratio,
